질문 등록 코드 수정 & Long 타입 변환

+++ src/main/java/com/hanaro/wouldyouhana/dto/CategoryDTO.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class CategoryDTO {
 
-    private Integer category_id;
+    private Long category_id;
 
     private String name;
